试卷批改

+++ OnLineExam/src/main/java/com/etc/exam/dao/TestPaperMapper.java
@@ -3,13 +3,13 @@ package com.etc.exam.dao;
 import com.etc.exam.entity.TestPaper;
 
 public interface TestPaperMapper {
-    int deleteByPrimaryKey(Short TESTPAPERID);
+    int deleteByPrimaryKey(int TESTPAPERID);
 
     int insert(TestPaper record);
 
     int insertSelective(TestPaper record);
 
-    TestPaper selectByPrimaryKey(Short TESTPAPERID);
+    TestPaper selectByPrimaryKey(int TESTPAPERID);
 
     int updateByPrimaryKeySelective(TestPaper record);
